WIP BER encoder
1. Length section
2. Value section
3. Change in DataType - typeID to unsigned long
4. Regex fixes
5.

+++ SNMPServer/Identifier.h
@@ -17,7 +17,7 @@ public:
 
 	 void setClass(string pClass);
 	 void setComplexity(string pComplexity);
-	 void setTag(unsigned int pTag);
+	 void setTag(unsigned long pTag);
 	 void setTagFromType(string pType);
 };
